% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.power_by_samplesize.R
\name{plot.power_by_samplesize}
\alias{plot.power_by_samplesize}
\title{Plot Power Results}
\usage{
\method{plot}{power_by_samplesize}(x, crit_power = NULL, plot_dir = NULL,
  ...)
}
\arguments{
\item{x}{list, results of a power function}

\item{crit_power}{numeric, critical power value one is looking for (adds a line
and the critical sample size to the plot)}

\item{plot_dir}{character, name of an existing directory where the power plots
should be saved to (leaving it  NULL will plot to
the default device)}

\item{...}{additional parameters, not used at the moment}
}
\description{
Plots the power-by-samplesize data.frames returned by the power functions.
}
\seealso{
\code{\link{boot.power.anova}}
}
